+++ CHANGELOG.md
@@ -7,3 +7,4 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 ### Added
  - Added abstractions for DmMd6XN switchers
+ - Adding DmMd6X1 adapter